### 2024-02-01 #6312 @editorialbot
[PRE REVIEW]: 21cmSense v2: A modular, open-source 21 cm sensitivity calculator
**Submitting author:** @steven-murray (<a href="http://orcid.org/0000-0003-3059-3823">Steven Murray</a>)
**Repository:** https://github.com/rasg-affiliates/21cmSense
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.0.0
