### 2024-02-01 4084e46
docs: add missing doi for pritchard
