+++ .github/workflows/release-draft.yml
@@ -18,6 +18,6 @@ jobs:
           python-version: "3.10"
 
       - name: Publish the release notes
-        uses: release-drafter/release-drafter@v5.25.0
+        uses: release-drafter/release-drafter@v6.0.0
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
