+++ .github/workflows/testsuite.yaml
@@ -22,7 +22,7 @@ jobs:
           fetch-depth: 1
 
       - name: Setup Miniconda
-        uses: conda-incubator/setup-miniconda@v3.0.1
+        uses: conda-incubator/setup-miniconda@v3.0.3
         with:
           auto-update-conda: true
           mamba-version: "*"
