### 2024-03-09 #6312 @arfon
@Hazboun6 @JulienPeloton @zonca :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is 21cmSense v2: A modular, open-source 21 cm sensitivity calculator

Based on your experience, and past submissions as authors here @ JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
