### 2024-03-18 #6501 @mattpitkin
Overall, the package is really well documented and the paper is nice. Well done! I just have some very minor comments at the moment:

* on the [contributing](https://21cmsense.readthedocs.io/en/latest/contributing.html) docs page, can you fix the links for bug reports/issues/etc.
* can you note that in order to run the tests you should pip install the package with `pip install .[test]` to get the required dependencies for pytest.
* in the paper, can you define acronyms on their first use (for [UV](https://en.wikipedia.org/wiki/UV_mapping) this might require a sentence stating what it is).
