+++ .pre-commit-config.yaml
@@ -54,7 +54,7 @@ repos:
     - id: rst-backticks
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v3.15.1
+  rev: v3.15.2
   hooks:
   -   id: pyupgrade
       args: [--py38-plus]
