### 2024-04-05 #6501 @htjb
Hi @steven-murray, sorry for the delay in my response. Overall the package is well written, well documented, and the tutorials are very helpful!  I particularly appreciate the detailed discussion of how the algorithm works (often something that is missing from scientific packages). I will definitely be investigating `commitizen` a bit more. My main comments are really on the paper, although these are not essential for publication in my mind.

At the moment the paper is very high level and I think you really need to understand the intricacies of the field to parse it. Although this is okay since the package is field specific, I think it could be helpful to expand the statement of need to explain in a few sentences why the 21cm line is important I.e. it can tell us about the properties of the first stars, structure formation and cosmology etc.

I think it would also be helpful to add an example of the kind of thing that 21cmSense can do into the paper as a visual reference for people. Maybe even just the graphs from the getting started tutorial showing the sensitivity of the HERA array relative to a realisation of the 21cm power spectrum from 21cmFAST and accompanied by a plot of the baseline positions.

Can also add a link to the relevant documentation on contributions into the repos README to help people find this.

Otherwise, super helpful package!
