+++ py21cmsense/data/antpos/layout_ecef_SKA_core.txt
@@ -1,224 +1,0 @@
--2.565031668943944853e+06 5.085463458483061753e+06 -2.861102804828245193e+06
--2.564918684477661736e+06 5.085647019469199702e+06 -2.860879316518400796e+06
--2.565009870781918522e+06 5.085524657505567186e+06 -2.861014164420139510e+06
--2.564822515250968281e+06 5.085621393657982349e+06 -2.861010203875160776e+06
--2.564985024936894886e+06 5.085573078753847629e+06 -2.860950795535154175e+06
--2.564972417855422478e+06 5.085667662456432357e+06 -2.860795012193705421e+06
--2.564951001280283090e+06 5.085620146294400096e+06 -2.860897987861589063e+06
--2.564938182463748381e+06 5.085569466665185988e+06 -2.860998888024772517e+06
--2.565066833286866546e+06 5.085560122163771652e+06 -2.860900816850310192e+06
--2.564829991432369221e+06 5.085673272313933820e+06 -2.860911944199128076e+06
--2.564876341431571636e+06 5.085650644279894419e+06 -2.860910624005768914e+06
--2.565126978715419304e+06 5.085566528603579849e+06 -2.860835938532195520e+06
--2.564930973128353246e+06 5.085541698887029663e+06 -2.861054335584214423e+06
--2.564971092680858914e+06 5.085535350281274877e+06 -2.861029817989191972e+06
--2.564916939598376397e+06 5.085609874731803313e+06 -2.860946457771202084e+06
--2.564768389477732591e+06 5.085679131972715259e+06 -2.860956453485653270e+06
--2.565008288537031040e+06 5.085602361207913607e+06 -2.860878373520457186e+06
--2.565109429846429266e+06 5.085541841363469139e+06 -2.860895158872165252e+06
--2.564900616675513797e+06 5.085584247260901146e+06 -2.861006243328803685e+06
--2.564995459376724903e+06 5.085549978021821938e+06 -2.860982291423859075e+06
--2.565106131512906868e+06 5.085515092648915946e+06 -2.860945326180333272e+06
--2.565116283204396721e+06 5.085472907087035477e+06 -2.861010769667385146e+06
--2.565003668868971057e+06 5.085502256333046593e+06 -2.861059239096883219e+06
--2.564938954325326718e+06 5.085665314435349777e+06 -2.860828960318984464e+06
--2.565081561241729185e+06 5.085596058822964318e+06 -2.860824245307636447e+06
--2.564905474216336384e+06 5.085696618846840225e+06 -2.860803499234514311e+06
--2.564990080823199823e+06 5.085631945202350616e+06 -2.860842350940564647e+06
--2.564942083024833817e+06 5.085496359061223455e+06 -2.861124493333651219e+06
--2.564881660671466496e+06 5.085558451626865193e+06 -2.861068668923002668e+06
--2.565042548196761869e+06 5.085602918977120891e+06 -2.860846877344298642e+06
--2.565036672972853296e+06 5.085498641497836448e+06 -2.861036230288373306e+06
--2.565057130032367073e+06 5.085628461376357824e+06 -2.860788788359759841e+06
--2.564858744279741775e+06 5.085545014039921574e+06 -2.861112800405434798e+06
--2.565148825095360167e+06 5.085520581933998503e+06 -2.860897610663039610e+06
--2.564860635809483472e+06 5.085592554878111929e+06 -2.861027177629656158e+06
--2.564796378190475516e+06 5.085601569434136152e+06 -2.861068480326557066e+06
--2.565027086224599741e+06 5.085550368900774978e+06 -2.860953435912403744e+06
--2.564842184046437033e+06 5.085638497833898291e+06 -2.860962488629752304e+06
[…]
