### 2024-04-05 5bd76c6
ci: don't test deploy to pypi, just check build
