### 2024-04-05 5d6d7d7
ci: test py311 and 312
