+++ .github/workflows/testsuite.yaml
@@ -15,7 +15,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: [3.9, "3.10"]
+        python-version: [3.9, "3.10", "3.11", "3.12"]
     steps:
       - uses: actions/checkout@master
         with:
@@ -44,7 +44,7 @@ jobs:
 
       - name: Run Tests
         run: |
-          python -m pytest --cov=py21cmsense --cov-config=.coveragerc --cov-report xml:./coverage.xml --junitxml=test-reports/xunit.xml
+          python -m pytest --cov=py21cmsense --cov-config=.coveragerc --cov-report xml:./coverage.xml
 
       - uses: codecov/codecov-action@v3
         if: success()
