### 2024-04-05 5a560a5
ci: test py311 and 312
