### 2024-04-05 018dec0
docs: udpate the readme to mention how to test the repo
