### 2024-04-05 8ed6fc2
docs: update joss paper reference to UV plane
