### 2024-04-08 acd0f69
docs: update joss paper to include plots
