+++ joss-paper/paper.bib
@@ -299,3 +299,41 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240104174S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Liu2020,
+       author = {{Liu}, Adrian and {Shaw}, J. Richard},
+        title = "{Data Analysis for Precision 21 cm Cosmology}",
+      journal = {\pasp},
+     keywords = {dark ages, reionization, first stars, methods: statistical, techniques: interferometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2020,
+        month = jun,
+       volume = {132},
+       number = {1012},
+          eid = {062001},
+        pages = {062001},
+          doi = {10.1088/1538-3873/ab5bfd},
+archivePrefix = {arXiv},
+       eprint = {1907.08211},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020PASP..132f2001L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{jwst,
+       author = {{Castellano}, Marco and {Fontana}, Adriano and {Treu}, Tommaso and {Santini}, Paola and {Merlin}, Emiliano and {Leethochawalit}, Nicha and {Trenti}, Michele and {Vanzella}, Eros and {Mestric}, Uros and {Bonchi}, Andrea and {Belfiori}, Davide and {Nonino}, Mario and {Paris}, Diego and {Polenta}, Gianluca and {Roberts-Borsani}, Guido and {Boyett}, Kristan and {Brada{\v{c}}}, Maru{\v{s}}a and {Calabr{\`o}}, Antonello and {Glazebrook}, Karl and {Grillo}, Claudio and {Mascia}, Sara and {Mason}, Charlotte and {Mercurio}, Amata and {Morishita}, Takahiro and {Nanayakkara}, Themiya and {Pentericci}, Laura and {Rosati}, Piero and {Vulcani}, Benedetta and {Wang}, Xin and {Yang}, Lilan},
+        title = "{Early Results from GLASS-JWST. III. Galaxy Candidates at z  9-15}",
+      journal = {\apjl},
+     keywords = {Reionization, 1383, Astrophysics - Astrophysics of Galaxies},
+         year = 2022,
+        month = oct,
+       volume = {938},
+       number = {2},
+          eid = {L15},
+        pages = {L15},
+          doi = {10.3847/2041-8213/ac94d0},
+archivePrefix = {arXiv},
+       eprint = {2207.09436},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJ...938L..15C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
