### 2024-04-08 1911d1e
docs: fix joss plots
