+++ joss-paper/paper.bib
@@ -337,3 +337,46 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJ...938L..15C},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Munoz2022,
+       author = {{Mu{\~n}oz}, Julian B. and {Qin}, Yuxiang and {Mesinger}, Andrei and {Murray}, Steven G. and {Greig}, Bradley and {Mason}, Charlotte},
+        title = "{The impact of the first galaxies on cosmic dawn and reionization}",
+      journal = {\mnras},
+     keywords = {galaxies: high-redshift, intergalactic medium, cosmology: theory, dark ages, reionization, first stars, diffuse radiation, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2022,
+        month = apr,
+       volume = {511},
+       number = {3},
+        pages = {3657-3681},
+          doi = {10.1093/mnras/stac185},
+archivePrefix = {arXiv},
+       eprint = {2110.13919},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022MNRAS.511.3657M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@article{Murray2020,
+  doi = {10.21105/joss.02582},
+  url = {https://doi.org/10.21105/joss.02582},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {54},
+  pages = {2582},
+  author = {Steven G. Murray and Bradley Greig and Andrei Mesinger and Julian B. Muñoz and Yuxiang Qin and Jaehong Park and Catherine A. Watkinson},
+  title = {21cmFAST v3: A Python-integrated C code for generating 3D realizations of the cosmic 21cm signal.},
+  journal = {Journal of Open Source Software}
+}
+
+@software{Mesinger2011,
+       author = {{Mesinger}, Andrei and {Furlanetto}, Steven and {Cen}, Renyue},
+        title = "{21cmFAST: A Fast, Semi-Numerical Simulation of the High-Redshift 21-cm Signal}",
+ howpublished = {Astrophysics Source Code Library, record ascl:1102.023},
+         year = 2011,
+        month = feb,
+          eid = {ascl:1102.023},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011ascl.soft02023M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
