+++ joss-paper/paper.md
@@ -97,13 +97,13 @@ An example of the predicted sensitivity of the HERA experiment after a year's ob
 at $z=8.5$ is shown in Figure \ref{sense}, corresponding to the sampling of the $(u,v)$-grid
 shown in Figure \ref{uvsampling}. The sensivity here is a signal-to-noise,
 assuming a signal magnitude computed using a semi-numerical model from the 21cmFAST
-code [@21cmfast], using parameters from [@Munoz22].
+code [@Mesinger2011; @Murray2020], using parameters from [@Munoz22].
 This figure also demonstrates that the new
 21cmSense is capable of producing sensitivity predictions in the cylindrically-averaged
 2D power spectrum space, which is helpful for upcoming experiments.
 
-![Sampling of the $(u,v)$-plane for the HERA experiment during a full year of observations.]{label="uvsampling"}(uv-sampling.png)
-![Predicted sensitivity of 1000 hours (one year) of HERA observations, as a function of perpendicular and line-of-sight fourier scale. The sensitivity is represented as the signal-to-noise on each $k$-mode, assuming a particular astrophysical model.]{label="sense"}(2dps.png)
+![Sampling of the $(u,v)$-plane for the HERA experiment during a full year of observations.](uv-sampling.png){label="uvsampling"}
+![Predicted sensitivity of 1000 hours (one year) of HERA observations, as a function of perpendicular and line-of-sight fourier scale. The sensitivity is represented as the signal-to-noise on each $k$-mode, assuming a particular astrophysical model.](2dps.png){label="sense"}
 
 # Statement of need
 
