+++ CONTRIBUTORS.rst
@@ -1,11 +1,8 @@
-Author
-======
+Authors
+=======
 
-`Jonathan Pober <https://github.com/jpober`_: jonathan_pober@brown.edu
-
-Contributors
-------------
-
+`Jonathan Pober <https://github.com/jpober>`_
+    - Original author of the code.
 - `Steven Murray <https://github.com/steven-murray>`_
     - Update to Python 3 and usability enhancements.
 - `Matthew Kolopanis <https://github.com/mkolopanis>`_
