### 2024-04-08 b28d8e3
docs: fix plot refs
