### 2024-04-08 c8ee2f0
docs: fix plot refs
