### 2024-04-08 086dfae
docs: fix plot refs
