### 2024-04-09 8ac238e
docs: use pdf plots?
