### 2024-04-15 #6501 @htjb
@steven-murray I agree with @mattpitkin, the latest version of the paper looks great, and again the code is well written and very useful. I did just notice that the superscript on the colorbar label in figure 2 hasn't rendered properly but other than that happy to sign off on this too.

@arfon My checklist is also complete, do we need to do anything else?
