### 2024-04-15 3d16f9b
fix: labels on 2D PS plots
