### 2024-04-15 #6501 @steven-murray
Thank you @mattpitkin and @htjb. There are a couple of open PRs on 21cmSense to address the issues you have helpfully pointed out (including fixing the colorbar label). Once reviewed things should be good to go.
