+++ .pre-commit-config.yaml
@@ -39,7 +39,7 @@ repos:
 
 
 - repo: https://github.com/psf/black-pre-commit-mirror
-  rev: 24.3.0
+  rev: 24.4.0
   hooks:
   - id: black
 
