### 2024-04-22 823bbb0
Update CONTRIBUTING.rst
Co-authored-by: Matthew Kolopanis <matthew.kolopanis@gmail.com>
