### 2024-04-22 62f224c
docs: fix Jonnies' author entry
