+++ CONTRIBUTORS.rst
@@ -1,7 +1,7 @@
 Authors
 =======
 
-`Jonathan Pober <https://github.com/jpober>`_
+- `Jonathan Pober <https://github.com/jpober>`_
     - Original author of the code.
 - `Steven Murray <https://github.com/steven-murray>`_
     - Update to Python 3 and usability enhancements.
