### 2024-04-22 83b7ae1
Doc updates (#114)
* maint: re-structure to src layout

* ci: don't test deploy to pypi, just check build

* ci: test py311 and 312

* docs: udpate the readme to mention how to test the repo

* docss: fix pypi badge

* docs: add contribution link into readme

* docs: fix links in contributing

* docs: fix authors

* Update CONTRIBUTING.rst

Co-authored-by: Matthew Kolopanis <matthew.kolopanis@gmail.com>

* fix: remove types module

* docs: fix Jonnies' author entry

---------

Co-authored-by: Matthew Kolopanis <matthew.kolopanis@gmail.com>
