### 2024-05-03 #6501 @steven-murray
@arfon yes --

https://github.com/rasg-affiliates/21cmSense/pull/117
https://github.com/rasg-affiliates/21cmSense/pull/116
https://github.com/rasg-affiliates/21cmSense/pull/114
https://github.com/rasg-affiliates/21cmSense/pull/113
