### 2024-05-08 #6501 @steven-murray
Hmmm, not sure what's going on here -- I updated the references without DOIs in https://github.com/rasg-affiliates/21cmSense/pull/122 (on the `joss-paper` branch) but it's complaining that this is the wrong branch. Does the paper now have to exist in main?
