### 2024-05-08 #6501 @arfon
:wave: @xuanxu – any chance you can take a quick look here? It seems like Editorialbot is struggling with the `joss-paper` branch but I'm not sure why?

> /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.9.1/lib/theoj/paper.rb:158:in `failure': Couldn't check the bibtex because branch name is incorrect: joss-paper (Theoj::Error)
