### 2024-05-08 #6501 @dfm
@steven-murray — There seem to be a few issues here. Can you take a stab at fixing them:

1. The error reported above (`citation Brietman2024 not found`) does seem to be right - that reference isn't in the `.bib` file.
2. I think that the syntax when referencing multiple sources needs to be changed to (note semicolon and space):
```diff
- [@cite1,@cite2]
+ [@cite1; @cite2]
```
3. A few of the journal references in the `.bib` file use AAS macros (e.g. `\mnras`). Those aren't available in our template, so please expand them appropriately.

Let me know when you've handled these last issues, and I'll take it from there!
