### 2024-05-09 2245342
docs: fix bib references
