+++ joss-paper/paper.bib
@@ -227,7 +227,7 @@
   author = {Pritchard, J. and Ichiki, K. and Mesinger, A. and Metcalf, R. B. and Pourtsidou, A. and Santos, M. and Abdalla, F. B. and Chang, T. C. and Chen, X. and Weller, J. and Zaroubi, S.},
   year = {2015},
   month = apr,
-  journal = {AASKA14},
+  journal = {Advancing Astrophysics with the Square Kilometre Array},
   pages = {12},
   url = {https://ui.adsabs.harvard.edu/abs/2015aska.confE..12P/abstract},
   urldate = {2019-11-01},
@@ -304,7 +304,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Liu2020,
        author = {{Liu}, Adrian and {Shaw}, J. Richard},
         title = "{Data Analysis for Precision 21 cm Cosmology}",
-      journal = {\pasp},
+      journal = {Publications of the Astronomical Society of the Pacific},
      keywords = {dark ages, reionization, first stars, methods: statistical, techniques: interferometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2020,
         month = jun,
@@ -323,7 +323,7 @@ archivePrefix = {arXiv},
 @ARTICLE{jwst,
        author = {{Castellano}, Marco and {Fontana}, Adriano and {Treu}, Tommaso and {Santini}, Paola and {Merlin}, Emiliano and {Leethochawalit}, Nicha and {Trenti}, Michele and {Vanzella}, Eros and {Mestric}, Uros and {Bonchi}, Andrea and {Belfiori}, Davide and {Nonino}, Mario and {Paris}, Diego and {Polenta}, Gianluca and {Roberts-Borsani}, Guido and {Boyett}, Kristan and {Brada{\v{c}}}, Maru{\v{s}}a and {Calabr{\`o}}, Antonello and {Glazebrook}, Karl and {Grillo}, Claudio and {Mascia}, Sara and {Mason}, Charlotte and {Mercurio}, Amata and {Morishita}, Takahiro and {Nanayakkara}, Themiya and {Pentericci}, Laura and {Rosati}, Piero and {Vulcani}, Benedetta and {Wang}, Xin and {Yang}, Lilan},
         title = "{Early Results from GLASS-JWST. III. Galaxy Candidates at z  9-15}",
-      journal = {\apjl},
+      journal = {The Astrophysical Journal Letters},
      keywords = {Reionization, 1383, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = oct,
@@ -342,7 +342,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Munoz2022,
        author = {{Mu{\~n}oz}, Julian B. and {Qin}, Yuxiang and {Mesinger}, Andrei and {Murray}, Steven G. and {Greig}, Bradley and {Mason}, Charlotte},
         title = "{The impact of the first galaxies on cosmic dawn and reionization}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {galaxies: high-redshift, intergalactic medium, cosmology: theory, dark ages, reionization, first stars, diffuse radiation, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = apr,
@@ -374,7 +374,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mesinger2011,
        author = {{Mesinger}, Andrei and {Furlanetto}, Steven and {Cen}, Renyue},
         title = "{21CMFAST: a fast, seminumerical simulation of the high-redshift 21-cm signal}",
[…]
