### 2024-05-09 #6501 @dfm
@steven-murray — Thanks! Looks like there's one (last! I hope) issue here. The problem is in the `.bib` file in one of the `abstract` entries. I'd recommend just deleting all the abstracts from the `.bib` - we don't need to track those. Thanks!!
