### 2024-05-09 1b556ee
docs: remove abstracts from bib file
