+++ joss-paper/paper.bib
@@ -1,5 +1,4 @@
-@article{
-    mwa,
+@article{mwa,
     title={The Murchison Widefield Array: The Square Kilometre Array Precursor at Low Radio Frequencies},
     volume={30},
     DOI={10.1017/pasa.2012.007},
@@ -32,7 +31,6 @@
   pages = {123},
   url = {http://adsabs.harvard.edu/abs/2018AJ....156..123A},
   urldate = {2020-09-28},
-  abstract = {The Astropy Project supports and fosters the development of open-source  and openly developed Python packages that provide commonly needed functionality to the astronomical community. A key element of the Astropy Project is the core package astropy, which serves as the foundation for more specialized projects and packages. In this article, we provide an overview of the organization of the Astropy project and summarize key features in the core package, as of the recent major release, version 2.0. We then describe the project infrastructure designed to facilitate and support development for a broader ecosystem of interoperable packages. We conclude with a future outlook of planned new features and directions for the broader Astropy Project. .},
   keywords = {methods: data analysis,methods: miscellaneous,methods: statistical,reference systems},
   annotation = {985 citations (Inspire/DOI) [2022-06-01]},
   doi = {10.3847/1538-3881/aabc4f}
@@ -49,7 +47,6 @@
   pages = {e026},
   url = {http://adsabs.harvard.edu/abs/2019PASA...36...26B},
   urldate = {2020-11-10},
-  abstract = {Epoch of Reionisation (EoR) data analysis requires unprecedented levels of accuracy in radio interferometer pipelines. We have developed an imaging power spectrum analysis to meet these requirements and generate robust 21 cm EoR measurements. In this work, we build a signal path framework to mathematically describe each step in the analysis, from data reduction in the Fast Holographic Deconvolution (FHD) package to power spectrum generation in the {$\varepsilon$}ppsilon package. In particular, we focus on the distinguishing characteristics of FHD/{$\varepsilon$}ppsilon: highly accurate spectral calibration, extensive data verification products, and end-to-end error propagation. We present our key data analysis products in detail to facilitate understanding of the prominent systematics in image-based power spectrum analyses. As a verification to our analysis, we also highlight a full-pipeline analysis simulation to demonstrate signal preservation and lack of signal loss. This careful treatment ensures that the FHD/{$\varepsilon$}ppsilon power spectrum pipeline can reduce radio interferometric data to produce credible 21 cm EoR measurements.},
   archiveprefix = {arxiv},
   keywords = {Astrophysics - Instrumentation and Methods for Astrophysics,cosmology: dark ages,data analysis,first stars,instrumentation,interferometers,interferometric,methods,reionisation,techniques},
   file = {C:\Users\steve\Zotero\storage\MB6SV4BT\1901.html},
@@ -68,7 +65,6 @@
   issn = {0035-8711},
   url = {https://ui.adsabs.harvard.edu/abs/2024MNRAS.527.9833B},
   urldate = {2024-01-18},
[…]
